requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-experimental==0.3.0
langchain-google-genai
google-generativeai
pandas
numpy
matplotlib
faiss-cpu
python-dotenv
streamlit
openpyxl